+++ src/Traits/EraserHelpers.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hith\LaravelEraser\Traits;
+namespace Eraser\Traits;
 
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;