+++ src/Traits/HasAutoEraser.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hith\LaravelEraser\Traits;
+namespace Eraser\Traits;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\Relation;