+++ tests/Pest.php
@@ -1,5 +1,7 @@
 <?php
 
-use Hith\LaravelEraser\Tests\TestCase;
+declare(strict_types=1);
+
+use Eraser\Tests\TestCase;
 
 pest()->extend(TestCase::class)->in('Feature', 'Unit');