+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [v1.0.0]
+## [2.0.0]
+### updates
+- Change namespace from Hith/LaravelEraser to Eraser
+- Change Manual Eraser array name from 'eraserRelationsToDelete' to 'erasable'
+
+## [1.0.0]
 ### Added
 - First version